+++ src/main/java/com/olap3/cubeexplorer/optimize/CostModel.java
@@ -2,5 +2,5 @@ package com.olap3.cubeexplorer.optimize;
 
 public interface CostModel {
     public void updateCost(TimeableOp operation);
-    public int correctedEstimate(TimeableOp operation);
+    public long correctedEstimate(TimeableOp operation);
 }
